Moved some files into better places
