+++ app/features/import/import-files.ts
@@ -1,19 +1,19 @@
 import {lakePath, workspacePath} from "app/router/utils/paths"
 import {Zealot} from "../../../zealot"
-import BrimApi from "../api"
-import brim from "../brim"
-import ingest from "../brim/ingest"
-import {IngestParams} from "../brim/ingest/getParams"
-import interop from "../brim/interop"
-import lib from "../lib"
-import Current from "../state/Current"
-import Handlers from "../state/Handlers"
-import {Handler} from "../state/Handlers/types"
-import Pools from "../state/Pools"
-import SystemTest from "../state/SystemTest"
-import Tabs from "../state/Tabs"
-import {Dispatch, Thunk} from "../state/types"
-import {getZealot} from "./getZealot"
+import BrimApi from "../../../src/js/api"
+import brim from "../../../src/js/brim"
+import ingest from "../../../src/js/brim/ingest"
+import {IngestParams} from "../../../src/js/brim/ingest/getParams"
+import interop from "../../../src/js/brim/interop"
+import lib from "../../../src/js/lib"
+import Current from "../../../src/js/state/Current"
+import Handlers from "../../../src/js/state/Handlers"
+import {Handler} from "../../../src/js/state/Handlers/types"
+import Pools from "../../../src/js/state/Pools"
+import SystemTest from "../../../src/js/state/SystemTest"
+import Tabs from "../../../src/js/state/Tabs"
+import {Dispatch, Thunk} from "../../../src/js/state/types"
+import {getZealot} from "../../../src/js/flows/getZealot"
 
 export default (files: File[]): Thunk<Promise<void>> => (
   dispatch,
